[DR-90] refactor: move EntityForm from internal form routes to global component
