+++ fe/cms/src/components/EntityForm/components/DisplayArtistStep.tsx
